requirements.txt
pydantic>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
asyncio>=3.4.3
pytest>=7.0.0
pytest-asyncio>=0.21.0
